Moved some classes into separate packages.
