+++ metals/src/main/scala/scala/meta/internal/bsp/BspServerConnectionFactory.scala
@@ -5,7 +5,6 @@ import scala.concurrent.Future
 import scala.concurrent.Promise
 import scala.util.Properties
 
-import scala.meta.internal.metals.BuildServerConnectionFactory
 import scala.meta.internal.metals.Cancelable
 import scala.meta.internal.metals.ClosableOutputStream
 import scala.meta.internal.metals.DismissedNotifications
@@ -14,9 +13,10 @@ import scala.meta.internal.metals.MetalsBuildClient
 import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.MetalsServerConfig
 import scala.meta.internal.metals.QuietInputStream
-import scala.meta.internal.metals.SocketConnection
 import scala.meta.internal.metals.TaskProgress
 import scala.meta.internal.metals.WorkDoneProgress
+import scala.meta.internal.metals.buildserver.BuildServerConnectionFactory
+import scala.meta.internal.metals.buildserver.SocketConnection
 import scala.meta.internal.metals.clients.language.ConfiguredLanguageClient
 import scala.meta.internal.mtags.URIEncoderDecoder
 import scala.meta.internal.process.SystemProcess
