+++ metals/src/main/scala/scala/meta/internal/metals/ImportedBuild.scala
@@ -8,6 +8,7 @@ import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 
 import scala.meta.internal.metals.MetalsEnrichments._
+import scala.meta.internal.metals.buildserver.BuildServerConnection
 
 import ch.epfl.scala.bsp4j._
