+++ metals/src/main/scala/scala/meta/internal/metals/JavaTarget.scala
@@ -2,6 +2,7 @@ package scala.meta.internal.metals
 
 import scala.meta.internal.builds.MillBuildTool
 import scala.meta.internal.metals.MetalsEnrichments._
+import scala.meta.internal.metals.buildserver.BuildServerConnection
 import scala.meta.internal.semver.SemVer
 import scala.meta.io.AbsolutePath
