+++ metals/src/main/scala/scala/meta/internal/metals/ScalaTarget.scala
@@ -12,6 +12,7 @@ import scala.meta.internal.io.PlatformFileIO
 import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.ScalaTarget.ExperimentalSyntaxRegex
 import scala.meta.internal.metals.ScalaTarget.KindProjectorRegex
+import scala.meta.internal.metals.buildserver.BuildServerConnection
 import scala.meta.internal.semver.SemVer
 import scala.meta.io.AbsolutePath
