+++ metals/src/main/scala/scala/meta/internal/metals/bloop/BloopServerConfigFactory.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.metals
+package scala.meta.internal.metals.bloop
 
 import java.lang.management.ManagementFactory
 import java.nio.file.Files
@@ -8,6 +8,7 @@ import scala.collection.concurrent.TrieMap
 import scala.util.Properties
 import scala.util.control.NonFatal
 
+import scala.meta.internal.metals.UserConfiguration
 import scala.meta.io.AbsolutePath
 
 import bloop.rifle.BloopRifleConfig
