+++ metals/src/main/scala/scala/meta/internal/metals/buildserver/BuildServerInitializeException.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.metals
+package scala.meta.internal.metals.buildserver
 
 class BuildServerInitializeException(e: Throwable)
     extends Exception("Failed to connect to build server", e)
